requirements.txt
python-nmap
scapy
tabulate
keyboard
